# Conflicts:
#	src/main/java/at/hannibal2/skyhanni/data/RenderData.kt

+++ src/main/java/at/hannibal2/skyhanni/utils/LorenzVec.kt
@@ -33,7 +33,7 @@ data class LorenzVec(
 
     constructor(x: Float, y: Float, z: Float) : this(x.toDouble(), y.toDouble(), z.toDouble())
 
-    fun toBlockPos(): BlockPos = BlockPos(x, y, z)
+    fun toBlockPos(): BlockPos = BlockPos(x.toInt(), y.toInt(), z.toInt())
 
     fun toVec3(): Vec3 = Vec3(x, y, z)
